Adding sprites to Animation screen.
Using the default sprite image that we have for omniview.
