+++ CMakeLists.txt
@@ -188,8 +188,105 @@ source_group(
 
 add_executable (
   CalChart
-  ${CalChartSources}
-  ${CalChartHeaders}
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationCanvas.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationCanvas.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationErrorsPanel.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationErrorsPanel.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationPanel.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationPanel.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationView.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/AnimationView.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CCOmniviewCanvas.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CCOmniviewCanvas.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartApp.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartApp.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartDoc.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartDoc.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartDocCommand.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartDocCommand.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartFrame.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartFrame.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartPreferences.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartPreferences.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartSizes.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartSizes.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartToolBar.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartToolBar.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartView.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CalChartView.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ColorPalette.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ColorPalette.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ColorSetupCanvas.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ColorSetupCanvas.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ColorSetupDialog.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ColorSetupDialog.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityBrowser.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityBrowser.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityBrowserPanel.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityBrowserPanel.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityComposerDialog.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityComposerDialog.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityEditorPopup.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ContinuityEditorPopup.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CustomListViewPanel.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/CustomListViewPanel.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/FieldCanvas.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/FieldCanvas.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/FieldControlsToolBar.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/FieldControlsToolBar.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/FieldThumbnailBrowser.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/FieldThumbnailBrowser.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesContCellSetup.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesContCellSetup.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesDrawingSetup.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesDrawingSetup.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesGeneralSetup.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesGeneralSetup.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesPSPrintingSetup.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesPSPrintingSetup.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesShowModeSetup.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesShowModeSetup.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesUtils.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PreferencesUtils.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PrintContinuityEditor.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/PrintContinuityEditor.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/TopFrame.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/TopFrame.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/background_image.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/background_image.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/basic_ui.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/basic_ui.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/cc_omniview_constants.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/confgr.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/confgr.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/cont_box_draw.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/cont_box_draw.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/draw.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/draw.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/draw_utils.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/e7_transition_solver_ui.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/e7_transition_solver_ui.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/field_canvas_shapes.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/field_canvas_shapes.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ghost_module.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ghost_module.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/linmath.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/mode_dialog.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/mode_dialog.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/mode_dialog_canvas.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/mode_dialog_canvas.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/platconf.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/precomp.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/precomp.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/print_ps_dialog.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/print_ps_dialog.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/setup_wizards.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/setup_wizards.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/show_ui.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/show_ui.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/single_instance_ipc.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/single_instance_ipc.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/ui_enums.h
   ${CalChartDocs}
   ${CalChartResources}
   "${CMAKE_CURRENT_SOURCE_DIR}/calchart.rc"
